Major Bug Fixes
Somehow didn't figure out that I should have been searching Proteins only even though the rest of the program was tailored to proteins, and even when it would search proteins, it would crash because I was searching the nucleotide database for proteins and even if it was used for a nucleotide, it would not work correctly. No idea why I thought the previous version approached functional.